Break out dates into a new module
