+++ cumulusci/tasks/bulkdata/tests/test_utils.py
@@ -1,4 +1,3 @@
-from datetime import datetime, date, timedelta
 import os
 import json
 import unittest
@@ -13,13 +12,8 @@ from cumulusci.utils import temporary_dir
 from cumulusci.tasks.bulkdata.utils import (
     create_table,
     generate_batches,
-    adjust_relative_dates,
-    get_relative_date_context,
-    datetime_from_salesforce,
-    salesforce_from_datetime,
 )
-from cumulusci.tasks.bulkdata.mapping_parser import parse_from_yaml, MappingStep
-from cumulusci.tasks.bulkdata.step import DataOperationType
+from cumulusci.tasks.bulkdata.mapping_parser import parse_from_yaml
 
 
 def create_db_file(filename):
@@ -193,132 +187,3 @@ class TestBatching(unittest.TestCase):
     def test_batching_with_remainder(self):
         batches = list(generate_batches(num_records=20, batch_size=7))
         assert batches == [(7, 0), (7, 1), (6, 2)]
-
-
-class TestRelativeDates:
-    def test_get_relative_date_context(self):
-        mapping = MappingStep(
-            sf_object="Account",
-            fields=["Some_Date__c", "Some_Datetime__c"],
-            anchor_date="2020-07-01",
-        )
-
-        org_config = mock.Mock()
-        org_config.salesforce_client.Account.describe.return_value = {
-            "fields": [
-                {"name": "Some_Date__c", "type": "date"},
-                {"name": "Some_Datetime__c", "type": "datetime"},
-                {"name": "Some_Bool__c", "type": "boolean"},
-            ]
-        }
-
-        assert get_relative_date_context(mapping, org_config) == (
-            [0],
-            [1],
-        )
-
-    def test_relative_dates(self):
-        mapping = MappingStep(
-            sf_object="Account", fields=["Some_Date__c"], anchor_date="2020-07-01"
-        )
-
-        target = date.today() + timedelta(days=7)
-        assert adjust_relative_dates(
-            mapping, ([0], []), ["2020-07-08"], DataOperationType.INSERT
-        ) == [target.isoformat()]
-
-        assert adjust_relative_dates(
-            mapping, ([0], []), ["2020-07-01"], DataOperationType.INSERT
-        ) == [date.today().isoformat()]
-
-        assert adjust_relative_dates(
-            mapping, ([0], []), [""], DataOperationType.INSERT
-        ) == [""]
-
-    def test_relative_dates__extract(self):
-        mapping = MappingStep(
-            sf_object="Account", fields=["Some_Date__c"], anchor_date="2020-07-01"
-        )
-
-        target = mapping.anchor_date + timedelta(days=7)
-        input_date = (date.today() + timedelta(days=7)).isoformat()
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([0], []),
-                ["001000000000000", input_date],
-                DataOperationType.QUERY,
-            )
-            == ["001000000000000", target.isoformat()]
-        )
-
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([0], []),
-                ["001000000000000", date.today().isoformat()],
-                DataOperationType.QUERY,
-            )
-            == ["001000000000000", mapping.anchor_date.isoformat()]
-        )
-
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([0], []),
-                ["001000000000000", ""],
-                DataOperationType.QUERY,
-            )
-            == ["001000000000000", ""]
-        )
-
-    def test_relative_datetimes(self):
-        mapping = MappingStep(
-            sf_object="Account", fields=["Some_Datetime__c"], anchor_date="2020-07-01"
-        )
-
-        input_dt = datetime_from_salesforce("2020-07-08T09:37:57.373+0000")
-        target = datetime.combine(date.today() + timedelta(days=7), input_dt.time())
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([], [0]),
-                [salesforce_from_datetime(input_dt)],
-                DataOperationType.INSERT,
-            )
-            == [salesforce_from_datetime(target)]
-        )
-
-        now = datetime.combine(mapping.anchor_date, datetime.now().time())
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([], [0]),
-                [salesforce_from_datetime(now)],
-                DataOperationType.INSERT,
-            )
-            == [salesforce_from_datetime(datetime.combine(date.today(), now.time()))]
-        )
-
-        assert adjust_relative_dates(
-            mapping, ([], [0]), [""], DataOperationType.INSERT
-        ) == [""]
-
-    def test_relative_datetimes_extract(self):
-        mapping = MappingStep(
-            sf_object="Account", fields=["Some_Datetime__c"], anchor_date="2020-07-01"
-        )
-
-        input_dt = datetime.now() + timedelta(days=7)
-        target = datetime.combine(
-            mapping.anchor_date + timedelta(days=7), input_dt.time()
-        )
-        assert (
-            adjust_relative_dates(
-                mapping,
-                ([], [0]),
-                ["001000000000000", salesforce_from_datetime(input_dt)],
-                DataOperationType.QUERY,
-            )
-            == ["001000000000000", salesforce_from_datetime(target)]
-        )
